major update
- ADDED: improved security of steam password handling
- ADDED: various minor improvements
- FIXED: mods where downloaded to the wrong folder after steamcmd update

+++ installer/adddelusr.sh
@@ -1,6 +1,11 @@
 #del users
-sudo deluser --remove-home $useradm
-sudo deluser --remove-home $userlnch
+if [ -d "/home/${usradm}" ]; then
+	sudo deluser --remove-home $useradm
+fi
+
+if [ -d "/home/${usrlnch}" ]; then
+	sudo deluser --remove-home $userlnch
+fi
 sudo groupdel $grpserver
 
 #mkusers
